requirements.txt
tornado==5.1.1
nose==1.3.7
fabric==2.4.0
gunicorn==19.9.0
psycopg2==2.7.2
tornado-sqlalchemy==0.5.0
alembic==1.0.1
beautifulsoup4==4.6.3
requests==2.20.0
mysqlclient==1.3.13
SQLAlchemy==1.2.12
pycrypto==2.6.1
